Create utils folder with sendEmail.js file
Create a new folder named utils in the root directory
Create a new folder named templates inside utils
Add sendEmail.js file inside utils folder
Implement Email class with newTransporter, send and sendWelcomeEmail methods
Add nodemailer, ejs and html-to-text dependencies
Set the view engine to ejs in index.js file
The sendEmail.js file implements an Email class with the newTransporter, send, and sendWelcomeEmail methods. These methods are used to send an email to the user. The newTransporter method creates a nodemailer transporter to send the email. The send method compiles an ejs template with the user's first name and sends an email to the user using the nodemailer transporter. The sendWelcomeEmail method sends a welcome email to the user.

Three dependencies, nodemailer, ejs and html-to-text have been added. The view engine has also been set to ejs in the index.js file.