Add forgot password functionality to auth controller
This commit adds the code for handling forgot password functionality in the auth controller. It extracts the email from the request body, checks if a user with the given email exists in the database, generates a password reset token, and sends a password reset email to the user.

It also adds a new method to the sendEmail class called sendPasswordReset() which sends a password reset email to the user.

Lastly, it creates a new email template called reset-password-email.ejs which contains the HTML markup for the password reset email.

Changes Made:

Added code for forgot password functionality in the auth controller
Added new method sendPasswordReset() in the sendEmail class
Created reset-password-email.ejs email template
File Changes:

Modified: authController.js
Modified: sendEmail.js
Added: reset-password-email.ejs