Added ResetPassword component to AllRoutes.jsx and created ResetPassword.jsx file
Added import { ResetPassword } from "./ResetPassword"; <Route path="/password_reset/edit" element={<ResetPassword />} /> in the client\src\Pages\AllRoutes.jsx

Created ResetPassword.jsx file in the client folder inside pages and added the ResetPassword component to it. The component displays a form that allows users to reset their password by providing a new password and confirming it. Upon submission, the component sends a patch request to the server with the user's new password and confirmation. If the request is successful, the user is redirected to the login page.