Update README.md file with Harvest Website Clone
This commit updates the README.md file with the Harvest Website Clone project. The project was built using HTML, CSS, JavaScript, React, Redux, Chakra UI, Node.js, Express.js, MongoDB, and Render. It is a time tracking and project management app that streamlines the process of logging team hours and storing them in one convenient location. The project features a responsive design that looks great on all devices, a user account authentication system with login and logout functionality, and various management systems including project, team, invoice, and password reset.

The README.md file includes a deployment link to the project and a list of the technologies used, along with images of the various screens and features. The images showcase the app's sleek design and user-friendly interface.

This commit also adds a new section to the README.md file, Glimpses of Harvest, which features images of the various screens and functionalities of the app.